requirements.txt
Flask==1.1.2
gunicorn==20.0.4
pytz
Pillow